requirements.txt
tqdm
matplotlib
circle_fit
scikit-image
pydicom
h5py
hdf5plugin
numpy
torch
scikit-learn
monai-weekly
tensorboard
